app/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
numpy==2.1.3
python-dateutil==2.9.0.post0
requests==2.32.3
langdetect==1.0.9
playwright==1.50.0
pytest==8.3.4
